Rename custom-errors.ts to custom_error.ts
